requirements.txt
certifi==2020.6.20
Pillow==8.4.0
